reorganize sensor files
