--- yahp/config.go
@@ -1,5 +1,0 @@
-package yahp
-
-type Config struct {
-
-}
