+++ yahp/sensor/install.go
@@ -4,7 +4,7 @@
 
 // +build windows
 
-package main
+package sensor
 
 import (
 	"fmt"
