+++ yahp/sensor/sensor_manager.go
@@ -1,4 +1,4 @@
-package yahp
+package sensor
 
 import (
 	"context"
